Share per unit minimum size logic to RN
