+++ packages/block-library/src/spacer/controls.js
@@ -2,10 +2,9 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { InspectorControls, useSetting } from '@wordpress/block-editor';
+import { InspectorControls } from '@wordpress/block-editor';
 import {
 	PanelBody,
-	__experimentalUseCustomUnits as useCustomUnits,
 	__experimentalUnitControl as UnitControl,
 	__experimentalParseQuantityAndUnitFromRawValue as parseQuantityAndUnitFromRawValue,
 } from '@wordpress/components';
@@ -14,7 +13,7 @@ import { useInstanceId } from '@wordpress/compose';
 /**
  * Internal dependencies
  */
-import { MIN_SPACER_SIZE } from './constants';
+import { useSpacerSettings } from './constants';
 
 function DimensionInput( { label, onChange, isResizing, value = '' } ) {
 	const inputId = useInstanceId( UnitControl, 'block-spacer-height-input' );
@@ -22,25 +21,7 @@ function DimensionInput( { label, onChange, isResizing, value = '' } ) {
 	// In most contexts the spacer size cannot meaningfully be set to a
 	// percentage, since this is relative to the parent container. This
 	// unit is disabled from the UI.
-	const availableUnitSettings = (
-		useSetting( 'spacing.units' ) || undefined
-	)?.filter( ( availableUnit ) => availableUnit !== '%' );
-
-	const units = useCustomUnits( {
-		availableUnits: availableUnitSettings || [
-			'px',
-			'em',
-			'rem',
-			'vw',
-			'vh',
-		],
-		defaultValues: { px: 100, em: 10, rem: 10, vw: 10, vh: 25 },
-	} );
-
-	// Finds the step for the current value’s unit for use as the min value.
-	const min =
-		units.find( ( { value: unit } ) => value.endsWith( unit ) )?.step ??
-		MIN_SPACER_SIZE;
+	const { units, min } = useSpacerSettings( value, '%' );
 
 	const handleOnChange = ( unprocessedValue ) => {
 		onChange( unprocessedValue );
